+++ mayan/apps/common/storages.py
@@ -1,11 +1,11 @@
 from __future__ import unicode_literals
 
-from django.utils.module_loading import import_string
+from mayan.apps.storage.utils import get_storage_subclass
 
 from .settings import (
     setting_shared_storage, setting_shared_storage_arguments
 )
 
-storage_sharedupload = import_string(
+storage_sharedupload = get_storage_subclass(
     dotted_path=setting_shared_storage.value
 )(**setting_shared_storage_arguments.value)